Remove call_symmetric from kernel trait.
This method is slower than `call` with multi-threading.

I attempted to parallelize call_symmetric_into by only running the parallelization on the upper triangle, then copying the upper triangle to the lower triangle.
However, the subsequent copy must be done syncronously, because we need to simultaneously read from one part of the matrix and write to another part.

This is not possible in Rust, because it would require holding both a mutable and immutable reference to the matrix.

We know it is safe in parallel, because we will only ever read from the upper half, and write to the lower half. Rust just does not have the necessary abstractions for this, as far as I can see.